fix: 📝 HPParameterType or Router Protocol 추가 ⛔️ TestTarget file 및 APIClient Singleton pattern 으로 수정
