+++ Core/HPNetwork/Sources/Base/Router.swift
@@ -1,37 +1,64 @@
 //
-//  TargetType.swift
+//  Router.swift
 //  HPNetwork
 //
-//  Created by 김진우 on 2023/05/22.
+//  Created by Kim dohyun on 2023/06/01.
 //
 
-import Alamofire
 import Foundation
 
-public protocol TargetType: URLRequestConvertible {
+import Alamofire
+
+
+public protocol Router: URLRequestConvertible {
+    
+    /// 공용 base URL
     var baseURL: String { get }
+    
+    /// HTTP Request Method
     var method: HTTPMethod { get }
+    
+    /// HTTP Request Path
     var path: String { get }
-    var parameters: RequestParams { get }
+    
+    /// HTTP Request Parameter
+    var parameters: HPParameterType { get }
+    
+    /// HTTP 공용 헤더 값
+    var headers: HTTPHeaders { get }
 }
 
-extension TargetType {
 
-    // URLRequestConvertible 구현
+
+extension Router {
+    
+    public var baseURL: String {
+        return "http://13.125.114.152:8080"
+    }
+    
+    public var parameters: HPParameterType {
+        return .none
+    }
+    
+    public var headers: HTTPHeaders {
+        return .default
+    }
+    
+    
     public func asURLRequest() throws -> URLRequest {
         let url = try baseURL.asURL()
         var urlRequest = try URLRequest(url: url.appendingPathComponent(path), method: method)
         urlRequest.setValue(ContentType.json.rawValue, forHTTPHeaderField: HTTPHeaderField.contentType.rawValue)
 
         switch parameters {
-        case .query(let request):
-            let params = request?.toDictionary() ?? [:]
+        case .query(let data):
+            let params = data?.toDictionary() ?? [:]
             let queryParams = params.map { URLQueryItem(name: $0.key, value: "\($0.value)") }
             var components = URLComponents(string: url.appendingPathComponent(path).absoluteString)
             components?.queryItems = queryParams
             urlRequest.url = components?.url
-        case .body(let request):
-            let params = request?.toDictionary() ?? [:]
+        case .body(let data):
+            let params = data?.toDictionary() ?? [:]
             urlRequest.httpBody = try JSONSerialization.data(withJSONObject: params, options: [])
         case .none:
             break
@@ -39,19 +66,7 @@ extension TargetType {
 
         return urlRequest
     }
+    
+    
 }
 
-public enum RequestParams {
-    case query(_ parameter: Encodable?)
-    case body(_ parameter: Encodable?)
-    case none
-}
-
-extension Encodable {
-    func toDictionary() -> [String: Any] {
-        guard let data = try? JSONEncoder().encode(self),
-              let jsonData = try? JSONSerialization.jsonObject(with: data),
-              let dictionaryData = jsonData as? [String: Any] else { return [:] }
-        return dictionaryData
-    }
-}
